--- persys-gateway/CHANGELOG.md
@@ -1,8 +1,0 @@
-# Change Log
-
-
-All notable changes to this project will be documented in this file.
-
-## [unreleased]
-
-* events manager is going to use etcd instead of mongoDB so we need to change api gateway accordingly
