+++ persys-gateway/cmd/main.go
@@ -14,11 +14,11 @@ import (
 
 	"github.com/gin-contrib/cors"
 	"github.com/gin-gonic/gin"
-	"github.com/persys-dev/persys-cloud/api-gateway/config"
-	"github.com/persys-dev/persys-cloud/api-gateway/controllers"
-	"github.com/persys-dev/persys-cloud/api-gateway/internal/middleware"
-	"github.com/persys-dev/persys-cloud/api-gateway/routes"
-	"github.com/persys-dev/persys-cloud/api-gateway/services"
+	"github.com/persys-dev/persys-cloud/persys-gateway/config"
+	"github.com/persys-dev/persys-cloud/persys-gateway/controllers"
+	"github.com/persys-dev/persys-cloud/persys-gateway/internal/middleware"
+	"github.com/persys-dev/persys-cloud/persys-gateway/routes"
+	"github.com/persys-dev/persys-cloud/persys-gateway/services"
 	ginprometheus "github.com/zsais/go-gin-prometheus"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
