+++ persys-gateway/controllers/auth.controller.go
@@ -10,10 +10,10 @@ import (
 	"github.com/dgrijalva/jwt-go/request"
 	"github.com/gin-gonic/gin"
 	"github.com/google/go-github/github"
-	"github.com/persys-dev/persys-cloud/api-gateway/config"
-	"github.com/persys-dev/persys-cloud/api-gateway/models"
-	"github.com/persys-dev/persys-cloud/api-gateway/services"
-	"github.com/persys-dev/persys-cloud/api-gateway/utils"
+	"github.com/persys-dev/persys-cloud/persys-gateway/config"
+	"github.com/persys-dev/persys-cloud/persys-gateway/models"
+	"github.com/persys-dev/persys-cloud/persys-gateway/services"
+	"github.com/persys-dev/persys-cloud/persys-gateway/utils"
 	"go.mongodb.org/mongo-driver/mongo"
 	"golang.org/x/oauth2"
 	oauth2gh "golang.org/x/oauth2/github"
