+++ persys-gateway/controllers/github.controller.go
@@ -8,8 +8,8 @@ import (
 	"github.com/gin-gonic/gin"
 	webhook "github.com/go-playground/webhooks/v6/github"
 	"github.com/google/go-github/github"
-	"github.com/persys-dev/persys-cloud/api-gateway/config"
-	"github.com/persys-dev/persys-cloud/api-gateway/services"
+	"github.com/persys-dev/persys-cloud/persys-gateway/config"
+	"github.com/persys-dev/persys-cloud/persys-gateway/services"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/mongo"
 	"golang.org/x/oauth2"
