+++ persys-gateway/services/auth.service.go
@@ -2,7 +2,7 @@ package services
 
 import (
 	"github.com/gin-gonic/gin"
-	"github.com/persys-dev/persys-cloud/api-gateway/models"
+	"github.com/persys-dev/persys-cloud/persys-gateway/models"
 )
 
 type AuthService interface {
