+++ persys-gateway/services/github.service.go
@@ -2,7 +2,7 @@ package services
 
 import (
 	"github.com/google/go-github/github"
-	"github.com/persys-dev/persys-cloud/api-gateway/models"
+	"github.com/persys-dev/persys-cloud/persys-gateway/models"
 	"go.mongodb.org/mongo-driver/bson"
 )
